Revert "[Rookie] 인덴테이션 및 변수명 수정"
