--- SwiftUI/CalculationUnitTest/CalculationUnitTest-Bridging-Header.h
@@ -1,4 +1,0 @@
-//
-//  Use this file to import your target's public headers that you would like to expose to Swift.
-//
-
