+++ SwiftUI/ReviewWQuestion-Calculator/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -72,11 +72,6 @@
     },
     {
       "idiom" : "ipad",
-      "scale" : "1x",
-      "size" : "76x76"
-    },
-    {
-      "idiom" : "ipad",
       "scale" : "2x",
       "size" : "76x76"
     },
