+++ SwiftUI/ReviewWQuestion-Calculator/ReviewWQuestion_CalculatorApp.swift
@@ -1,0 +1,15 @@
+//
+//  ReviewWQuestion_CalculatorApp.swift
+//  ReviewWQuestion-Calculator
+//
+
+import SwiftUI
+
+@main
+struct ReviewWQuestion_CalculatorApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
