--- SwiftUI/ReviewWQuestion-Calculator/System/CalculationApp.swift
@@ -1,17 +1,0 @@
-//
-//  ReviewWQuestion_CalculatorApp.swift
-//  ReviewWQuestion-Calculator
-//
-
-import SwiftUI
-
-@main
-struct CalculationApp: App {
-    @StateObject var calculationData = CalculationData()
-    var body: some Scene {
-        WindowGroup {
-            CalculationView()
-                .environmentObject(calculationData)
-        }
-    }
-}
