--- SwiftUI/ReviewWQuestion-Calculator/es.lproj/Localizable.strings
@@ -1,7 +1,0 @@
-/* 
-  Localizable.strings
-  ReviewWQuestion-Calculator
-
-  Created by Jisu Jang on 2022/06/01.
-  
-*/
