--- SwiftUI/ReviewWQuestion-Calculator/fr.lproj/Localizable.strings
@@ -1,8 +1,0 @@
-/* 
-  Localizable.strings
-  ReviewWQuestion-Calculator
-
-  Created by Jisu Jang on 2022/06/01.
-  
-*/
-"error" = "Erreur";
